Move calculateTgetSelector into lib/util.ts
